Added mouse button support for non-Windows platforms. > The only change needed here is the modification of the handleClick method signature of SkWindow. Occurances searched with 'grep "handleClick(" . -r' in the root directory.
On OSX, due to the one button nature of the platform, all clicks are done with the left button. This is the same function as before when only the left button was supported.

In XLib and SDL, mouse button numbers should align number for number. The
only fuzzy area is in XLib where Button2 is middle and Button3 is right.
This seems to be the concensus.

Also, added my name to AUTHORS